requirements.txt
langchain>=0.1.16,<0.2.0
langchain-community>=0.0.30  # Required for community integrations
langchain-core>=0.1.40,<0.2.0
langgraph>=0.0.40
crewai>=0.20.3
groq>=0.3.2
langchain-groq>=0.0.2
pydantic>=2.7.1
fastapi>=0.111.0
uvicorn>=0.29.0
sqlite-utils>=3.36
PyYAML>=6.0.1
httpx>=0.27.0
rich>=13.7.1
python-dotenv>=1.0.1
youtube-transcript-api>=0.4.5
sentence-transformers>=2.6.1
scikit-learn>=1.5.0
chromadb
sentence-transformers
